requirements.txt
gunicorn
altair==5.1.2
amqp==5.2.0
async-timeout==4.0.3
attrs==23.1.0
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ffmpeg-python==0.2.0
Flask==3.0.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kombu==5.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
prompt-toolkit==3.0.41
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
rich==13.7.0
rpds-py==0.13.1
six==1.16.0
smmap==5.0.1
streamlit==1.28.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
vine==5.1.0
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.10
Werkzeug==3.0.1
whispercpp @ git+https://github.com/stlukey/whispercpp.py@7af678159c29edb3bc2a51a72665073d58f2352f
zipp==3.17.0
